+++ src/main/java/racingcar/exception/car/InvalidNameException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.car;
 
 public class InvalidNameException extends IllegalArgumentException {
     public InvalidNameException(final String s) {
