+++ src/main/java/racingcar/exception/repeater/InvalidCountException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.repeater;
 
 public class InvalidCountException extends IllegalArgumentException {
     public InvalidCountException(final String s) {
