+++ src/main/java/racingcar/exception/round/InvalidRoundException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.round;
 
 public class InvalidRoundException extends IllegalArgumentException {
     public InvalidRoundException(final String s) {
