+++ src/main/java/racingcar/exception/splitter/InvalidDelimiterException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.splitter;
 
 public class InvalidDelimiterException extends IllegalArgumentException {
     public InvalidDelimiterException(final String s) {
