+++ src/main/java/racingcar/support/splitter/Splitter.java
@@ -1,6 +1,6 @@
 package racingcar.support.splitter;
 
-import racingcar.exception.InvalidDelimiterException;
+import racingcar.exception.splitter.InvalidDelimiterException;
 
 public class Splitter {
