+++ src/test/java/racingcar/model/game/round/RoundTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.exception.InvalidRoundException;
+import racingcar.exception.round.InvalidRoundException;
 
 @DisplayName("시도 횟수 테스트")
 class RoundTest {
